% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binfhe.R
\name{bin_decrypt}
\alias{bin_decrypt}
\title{Decrypt a Binary FHE ciphertext}
\usage{
bin_decrypt(ctx, sk, ct, p = 4L)
}
\arguments{
\item{ctx}{A BinFHE context}

\item{sk}{An LWEPrivateKey}

\item{ct}{An LWECiphertext}

\item{p}{Plaintext modulus (default: 4)}
}
\value{
Integer value
}
\description{
Decrypt a Binary FHE ciphertext
}
