% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binfhe.R
\name{bin_key_gen}
\alias{bin_key_gen}
\title{Generate BinFHE secret key}
\usage{
bin_key_gen(ctx)
}
\arguments{
\item{ctx}{A BinFHE context}
}
\value{
An LWEPrivateKey
}
\description{
Generate BinFHE secret key
}
