% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{deserialize_eval_keys}
\alias{deserialize_eval_keys}
\title{Deserialize evaluation keys from file}
\usage{
deserialize_eval_keys(
  filename,
  type = c("mult", "automorphism", "sum"),
  format = "binary"
)
}
\arguments{
\item{filename}{Path to serialized file}

\item{type}{"mult", "automorphism", or "sum"}

\item{format}{"binary" (default) or "json"}
}
\value{
TRUE on success (invisibly)
}
\description{
Deserialize evaluation keys from file
}
