% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{eval_add}
\alias{eval_add}
\title{Homomorphic addition}
\usage{
eval_add(x, y, ...)
}
\arguments{
\item{x, y}{Ciphertext, Plaintext, or numeric values}

\item{...}{Method-specific arguments}
}
\value{
A Ciphertext
}
\description{
Homomorphic addition
}
