% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ckks.R
\name{eval_bootstrap}
\alias{eval_bootstrap}
\title{Perform CKKS bootstrapping}
\usage{
eval_bootstrap(ct, num_iterations = 1L, precision = 0L)
}
\arguments{
\item{ct}{A Ciphertext}

\item{num_iterations}{Number of bootstrap iterations (default: 1)}

\item{precision}{Target precision (default: 0 = automatic)}
}
\value{
A refreshed Ciphertext
}
\description{
Refreshes the ciphertext to allow further computation.
}
