% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ckks.R
\name{eval_chebyshev_linear}
\alias{eval_chebyshev_linear}
\title{Evaluate a Chebyshev series via the linear evaluator}
\usage{
eval_chebyshev_linear(ct, coefficients, a, b)
}
\arguments{
\item{ct}{A Ciphertext}

\item{coefficients}{Numeric vector of Chebyshev coefficients}

\item{a}{Lower bound of the approximation interval}

\item{b}{Upper bound of the approximation interval}
}
\value{
A Ciphertext
}
\description{
Forces the "linear" Chebyshev evaluator regardless of degree.
Shallower circuit depth than Paterson-Stockmeyer but uses
more multiplications at high degree.
}
\seealso{
\code{\link[=eval_chebyshev]{eval_chebyshev()}}, \code{\link[=eval_chebyshev_ps]{eval_chebyshev_ps()}}
}
