% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ckks.R
\name{eval_chebyshev_ps}
\alias{eval_chebyshev_ps}
\title{Evaluate a Chebyshev series via the Paterson-Stockmeyer method}
\usage{
eval_chebyshev_ps(ct, coefficients, a, b)
}
\arguments{
\item{ct}{A Ciphertext}

\item{coefficients}{Numeric vector of Chebyshev coefficients}

\item{a}{Lower bound of the approximation interval}

\item{b}{Upper bound of the approximation interval}
}
\value{
A Ciphertext
}
\description{
Forces the Paterson-Stockmeyer Chebyshev evaluator regardless
of degree. Efficient for high-degree polynomials.
}
\seealso{
\code{\link[=eval_chebyshev]{eval_chebyshev()}}, \code{\link[=eval_chebyshev_linear]{eval_chebyshev_linear()}}
}
