% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{eval_mult}
\alias{eval_mult}
\title{Homomorphic multiplication}
\usage{
eval_mult(x, y, ...)
}
\arguments{
\item{x, y}{Ciphertext, Plaintext, or numeric values}

\item{...}{Method-specific arguments}
}
\value{
A Ciphertext
}
\description{
Homomorphic multiplication
}
