% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-eval-9111.R
\name{eval_mult_and_relinearize}
\alias{eval_mult_and_relinearize}
\title{Fused multiply-and-relinearize}
\usage{
eval_mult_and_relinearize(x, ...)
}
\arguments{
\item{x}{A \code{Ciphertext}.}

\item{...}{Method-specific arguments: \code{y} — a \code{Ciphertext} to
multiply \code{x} by in a fused multiply-and-relinearize step.}
}
\value{
A relinearized \code{Ciphertext}.
}
\description{
Equivalent to \code{relinearize(eval_mult_no_relin(x, y))} but
slightly more efficient in OpenFHE's implementation. Use this
at the end of a multiplication chain.
}
