% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{eval_negate}
\alias{eval_negate}
\title{Homomorphic negation}
\usage{
eval_negate(x, ...)
}
\arguments{
\item{x}{A Ciphertext}

\item{...}{Method-specific arguments}
}
\value{
A Ciphertext
}
\description{
Homomorphic negation
}
