% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-eval-9111.R
\name{eval_negate_in_place}
\alias{eval_negate_in_place}
\title{Homomorphic negation in place}
\usage{
eval_negate_in_place(x, ...)
}
\arguments{
\item{x}{A \code{Ciphertext} (modified in place).}

\item{...}{Reserved for future method-specific arguments.}
}
\value{
\code{x} invisibly.
}
\description{
Homomorphic negation in place
}
