% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binfhe.R
\name{eval_not}
\alias{eval_not}
\title{Evaluate NOT on an encrypted value}
\usage{
eval_not(ctx, ct)
}
\arguments{
\item{ctx}{A BinFHE context}

\item{ct}{An LWECiphertext}
}
\value{
An LWECiphertext
}
\description{
Evaluate NOT on an encrypted value
}
