% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-eval.R
\name{eval_rotate}
\alias{eval_rotate}
\title{Rotate ciphertext slots}
\usage{
eval_rotate(ct, ...)
}
\arguments{
\item{ct}{A Ciphertext}

\item{...}{Method-specific arguments (index)}
}
\value{
A Ciphertext
}
\description{
Rotate ciphertext slots
}
