% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ckks.R
\name{eval_sin}
\alias{eval_sin}
\title{Evaluate sine on a ciphertext (Chebyshev approximation)}
\usage{
eval_sin(ct, a, b, degree)
}
\arguments{
\item{ct}{A Ciphertext}

\item{a}{Lower bound of the approximation interval}

\item{b}{Upper bound of the approximation interval}

\item{degree}{Chebyshev polynomial degree}
}
\value{
A Ciphertext
}
\description{
Evaluate sine on a ciphertext (Chebyshev approximation)
}
