% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{eval_square}
\alias{eval_square}
\title{Homomorphic squaring}
\usage{
eval_square(x, ...)
}
\arguments{
\item{x}{A Ciphertext}

\item{...}{Method-specific arguments}
}
\value{
A Ciphertext
}
\description{
Homomorphic squaring
}
