% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-eval-9111.R
\name{eval_square_mutable}
\alias{eval_square_mutable}
\title{Homomorphic square, mutable variant}
\usage{
eval_square_mutable(x, ...)
}
\arguments{
\item{x}{A \code{Ciphertext} (may be modified internally).}

\item{...}{Reserved for future method-specific arguments.}
}
\value{
A new \code{Ciphertext} holding \code{x * x}.
}
\description{
Homomorphic square, mutable variant
}
