% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{eval_sub}
\alias{eval_sub}
\title{Homomorphic subtraction}
\usage{
eval_sub(x, y, ...)
}
\arguments{
\item{x, y}{Ciphertext, Plaintext, or numeric values}

\item{...}{Method-specific arguments}
}
\value{
A Ciphertext
}
\description{
Homomorphic subtraction
}
