% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-eval.R
\name{eval_sum}
\alias{eval_sum}
\title{Sum all slots in a ciphertext}
\usage{
eval_sum(ct, ...)
}
\arguments{
\item{ct}{A Ciphertext}

\item{...}{Method-specific arguments (batch_size)}
}
\value{
A Ciphertext
}
\description{
Sum all slots in a ciphertext
}
