% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{fhe_deserialize}
\alias{fhe_deserialize}
\title{Deserialize an OpenFHE object from file}
\usage{
fhe_deserialize(
  filename,
  type = c("CryptoContext", "PublicKey", "PrivateKey", "Ciphertext"),
  format = "binary"
)
}
\arguments{
\item{filename}{Path to serialized file}

\item{type}{One of "CryptoContext", "PublicKey", "PrivateKey", "Ciphertext"}

\item{format}{"binary" (default) or "json"}
}
\value{
The deserialized object
}
\description{
Deserialize an OpenFHE object from file
}
