% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{fhe_serialize}
\alias{fhe_serialize}
\title{Serialize an OpenFHE object to file}
\usage{
fhe_serialize(x, ...)
}
\arguments{
\item{x}{An OpenFHE object (CryptoContext, PublicKey, PrivateKey, Ciphertext)}

\item{...}{Method-specific arguments (filename, format)}
}
\value{
TRUE on success (invisibly)
}
\description{
Serialize an OpenFHE object to file
}
