% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ckks.R
\name{get_bootstrap_depth}
\alias{get_bootstrap_depth}
\title{Get the required multiplicative depth for CKKS bootstrapping}
\usage{
get_bootstrap_depth(level_budget, secret_key_dist = 1L)
}
\arguments{
\item{level_budget}{Integer vector of length 2}

\item{secret_key_dist}{Secret key distribution (default: UNIFORM_TERNARY = 1)}
}
\value{
Integer: required depth
}
\description{
Get the required multiplicative depth for CKKS bootstrapping
}
