% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-eval.R
\name{get_ckks_boot_correction_factor}
\alias{get_ckks_boot_correction_factor}
\title{Get the CKKS bootstrap correction factor}
\usage{
get_ckks_boot_correction_factor(cc)
}
\arguments{
\item{cc}{A \code{CryptoContext}.}
}
\value{
Integer; the current correction factor.
}
\description{
Reads the current correction factor the scheme uses during
the bootstrap \code{EvalModReduceInternal} step. Companion of
\code{\link[=set_ckks_boot_correction_factor]{set_ckks_boot_correction_factor()}}. Changes here affect
all subsequent bootstrap operations on this \code{CryptoContext}
until another call to \code{\link[=set_ckks_boot_correction_factor]{set_ckks_boot_correction_factor()}}.
}
\seealso{
\code{\link[=set_ckks_boot_correction_factor]{set_ckks_boot_correction_factor()}}, \code{\link[=eval_bootstrap_setup]{eval_bootstrap_setup()}}
}
