% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptocontext-getters.R
\name{get_cyclotomic_order}
\alias{get_cyclotomic_order}
\title{Cyclotomic order of a CryptoContext}
\usage{
get_cyclotomic_order(cc, ...)
}
\arguments{
\item{cc}{A \code{CryptoContext}.}

\item{...}{Reserved for future method-specific arguments.}
}
\value{
Integer.
}
\description{
Integer \code{m} such that the underlying polynomial ring is
\code{Z[x]/(x^n + 1)} with \code{n = m/2} (the ring dimension). Always
\code{2 * ring_dimension(cc)} for power-of-two cyclotomics.
}
