% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptocontext-getters.R
\name{get_element_params}
\alias{get_element_params}
\title{Element parameters of a CryptoContext}
\usage{
get_element_params(cc, ...)
}
\arguments{
\item{cc}{A \code{CryptoContext}.}

\item{...}{Reserved for future method-specific arguments.}
}
\value{
An \code{ElementParams} S7 object.
}
\description{
Returns the \code{ElementParams} S7 object wrapping
\verb{std::shared_ptr<typename DCRTPoly::Params>}. This is the
object that can be passed as the \code{params} argument to
\code{make_ckks_packed_plaintext()} to build a plaintext against a
specific parameter set rather than the context default.
}
\details{
This is the first R-side way to obtain a non-default
\code{ElementParams} (the class itself has existed as a scaffold
but had no constructor path until now).
}
