% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binfhe.R
\name{get_max_plaintext_space}
\alias{get_max_plaintext_space}
\title{Maximum supported plaintext space for functional bootstrapping}
\usage{
get_max_plaintext_space(ctx)
}
\arguments{
\item{ctx}{A BinFHE context built with \code{arb_func = TRUE}}
}
\value{
A numeric scalar (q / (2 * beta))
}
\description{
Maximum supported plaintext space for functional bootstrapping
}
