% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_native_int}
\alias{get_native_int}
\title{Get the native integer size of the OpenFHE build}
\usage{
get_native_int()
}
\value{
integer
}
\description{
Returns 64 or 128 depending on how OpenFHE was compiled.
}
