% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaintext.R
\name{get_packed_value}
\alias{get_packed_value}
\title{Get packed integer values from a plaintext}
\usage{
get_packed_value(pt)
}
\arguments{
\item{pt}{A Plaintext}
}
\value{
Integer vector
}
\description{
Get packed integer values from a plaintext
}
