% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfhe-object.R
\name{get_ptr}
\alias{get_ptr}
\title{Extract pointer with validation}
\usage{
get_ptr(x)
}
\arguments{
\item{x}{An OpenFHEObject}
}
\value{
The external pointer
}
\description{
Extract pointer with validation
}
\keyword{internal}
