% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive-bootstrap.R
\name{int_boot_adjust_scale}
\alias{int_boot_adjust_scale}
\title{Prepare a ciphertext for interactive bootstrap}
\usage{
int_boot_adjust_scale(ct)
}
\arguments{
\item{ct}{A \code{Ciphertext}.}
}
\value{
A \code{Ciphertext} ready for \code{\link[=int_boot_decrypt]{int_boot_decrypt()}}.
}
\description{
Adjusts a ciphertext's scale to meet the scheme's
requirements before entering the interactive bootstrap
protocol. Typically called before \code{\link[=int_boot_decrypt]{int_boot_decrypt()}}.
}
