% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive-bootstrap.R
\name{int_mp_boot_adjust_scale}
\alias{int_mp_boot_adjust_scale}
\title{Prepare a ciphertext for multi-party interactive bootstrap}
\usage{
int_mp_boot_adjust_scale(ct)
}
\arguments{
\item{ct}{A \code{Ciphertext}.}
}
\value{
A \code{Ciphertext} ready for the multi-party bootstrap
protocol.
}
\description{
Multi-party analogue of \code{\link[=int_boot_adjust_scale]{int_boot_adjust_scale()}}. Adjusts
the ciphertext's scale before entering the distributed
bootstrap protocol.
}
