% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-eval-9111.R
\name{mod_reduce_in_place}
\alias{mod_reduce_in_place}
\title{Reduce the modulus chain by one level, in place}
\usage{
mod_reduce_in_place(x, ...)
}
\arguments{
\item{x}{A \code{Ciphertext} (modified in place).}

\item{...}{Reserved for future method-specific arguments.}
}
\value{
\code{x} invisibly.
}
\description{
Reduce the modulus chain by one level, in place
}
