% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval-key-map.R
\name{multi_add_eval_sum_keys}
\alias{multi_add_eval_sum_keys}
\title{Combine two sum-key map shares into a joint sum-key map}
\usage{
multi_add_eval_sum_keys(cc, eval_key_map1, eval_key_map2, key_tag = "")
}
\arguments{
\item{cc}{A \code{CryptoContext}.}

\item{eval_key_map1, eval_key_map2}{\code{EvalKeyMap} shares from
two parties.}

\item{key_tag}{Character; default \code{""}.}
}
\value{
A combined \code{EvalKeyMap} suitable for insertion into
the cc registry via \code{\link[=insert_eval_sum_key]{insert_eval_sum_key()}}.
}
\description{
Combine two sum-key map shares into a joint sum-key map
}
