% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiparty.R
\name{multiparty_decrypt_main}
\alias{multiparty_decrypt_main}
\title{Non-lead party's partial decryption}
\usage{
multiparty_decrypt_main(cc, sk, ct)
}
\arguments{
\item{cc}{A CryptoContext}

\item{sk}{This party's PrivateKey}

\item{ct}{A Ciphertext or a list of Ciphertexts}
}
\value{
A partially decrypted Ciphertext or list of
Ciphertexts, mirroring the input shape.
}
\description{
Other parties call this after the lead. Accepts either a
single \code{Ciphertext} or a list of \code{Ciphertext} objects with the
same semantics as \code{\link[=multiparty_decrypt_lead]{multiparty_decrypt_lead()}}.
}
