% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfhe-object.R
\name{ptr_is_valid}
\alias{ptr_is_valid}
\title{Check if an external pointer is valid (non-NULL)}
\usage{
ptr_is_valid(x)
}
\arguments{
\item{x}{An OpenFHEObject}
}
\value{
logical
}
\description{
Check if an external pointer is valid (non-NULL)
}
\keyword{internal}
