% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{ring_dimension}
\alias{ring_dimension}
\title{Ring dimension of a CryptoContext}
\usage{
ring_dimension(cc, ...)
}
\arguments{
\item{cc}{A CryptoContext}

\item{...}{Reserved for future method-specific arguments}
}
\value{
Integer ring dimension
}
\description{
Returns N, the cyclotomic ring dimension. The cyclotomic order
M used by \code{eval_fast_rotation()} is \code{2 * N}.
}
