% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{serialize_eval_keys}
\alias{serialize_eval_keys}
\title{Serialize evaluation keys to file}
\usage{
serialize_eval_keys(
  filename,
  type = c("mult", "automorphism", "sum"),
  format = "binary",
  key_tag = ""
)
}
\arguments{
\item{filename}{Path to output file}

\item{type}{"mult" or "automorphism"}

\item{format}{"binary" (default) or "json"}

\item{key_tag}{Key tag (default: "")}
}
\value{
TRUE on success (invisibly)
}
\description{
Serialize evaluation keys to file
}
