% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptocontext-getters.R
\name{set_key_gen_level}
\alias{set_key_gen_level}
\title{Set the key-generation level of a CryptoContext}
\usage{
set_key_gen_level(cc, ...)
}
\arguments{
\item{cc}{A \code{CryptoContext}.}

\item{...}{Method-specific arguments; the \code{level} integer is
passed here.}
}
\value{
\code{cc} invisibly.
}
\description{
Set the key-generation level of a CryptoContext
}
