% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{with_fhe_context}
\alias{with_fhe_context}
\title{Execute code with automatic cleanup of FHE state}
\usage{
with_fhe_context(expr)
}
\arguments{
\item{expr}{Expression to evaluate}
}
\value{
Result of expr
}
\description{
Clears eval keys and releases contexts on exit (even on error).
}
