% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{BinFHEMethod}
\alias{BinFHEMethod}
\title{Binary FHE Methods
Source: binfhe/binfhe-constants.h enum BINFHE_METHOD}
\usage{
BinFHEMethod
}
\description{
Binary FHE Methods
Source: binfhe/binfhe-constants.h enum BINFHE_METHOD
}
