% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{BinFHEOutput}
\alias{BinFHEOutput}
\title{Binary FHE Output Types
Source: binfhe/binfhe-constants.h enum BINFHE_OUTPUT}
\usage{
BinFHEOutput
}
\description{
Binary FHE Output Types
Source: binfhe/binfhe-constants.h enum BINFHE_OUTPUT
}
