% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{BinFHEParamSet}
\alias{BinFHEParamSet}
\title{Binary FHE Parameter Sets
Source: binfhe/binfhe-constants.h enum BINFHE_PARAMSET (sequential from 0)}
\usage{
BinFHEParamSet
}
\description{
Binary FHE Parameter Sets
Source: binfhe/binfhe-constants.h enum BINFHE_PARAMSET (sequential from 0)
}
