% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{BinGate}
\alias{BinGate}
\title{Binary Gate Types
Source: binfhe/binfhe-constants.h enum BINGATE (sequential from 0)}
\usage{
BinGate
}
\description{
Binary Gate Types
Source: binfhe/binfhe-constants.h enum BINGATE (sequential from 0)
}
