% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{clear_fhe_state}
\alias{clear_fhe_state}
\title{Clear cached evaluation keys and contexts}
\usage{
clear_fhe_state(what = c("mult_keys", "automorphism_keys", "contexts"))
}
\arguments{
\item{what}{Character vector: subset of "mult_keys", "automorphism_keys", "contexts"}
}
\description{
Clear cached evaluation keys and contexts
}
