% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-encrypt.R
\name{decrypt}
\alias{decrypt}
\title{Decrypt a ciphertext}
\usage{
decrypt(ct, key, ...)
}
\arguments{
\item{ct}{A Ciphertext}

\item{key}{A PrivateKey}

\item{...}{Additional arguments (cc = CryptoContext)}
}
\value{
A Plaintext
}
\description{
Decrypt a ciphertext
}
