% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{DecryptionNoiseMode}
\alias{DecryptionNoiseMode}
\title{Decryption Noise Mode
Source: pke/constants-defs.h enum DecryptionNoiseMode}
\usage{
DecryptionNoiseMode
}
\description{
Decryption Noise Mode
Source: pke/constants-defs.h enum DecryptionNoiseMode
}
