% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{DistributionType}
\alias{DistributionType}
\title{Distribution Type (lattice parameters)
Source: core/lattice/stdlatticeparms.h enum DistributionType}
\usage{
DistributionType
}
\description{
Distribution Type (lattice parameters)
Source: core/lattice/stdlatticeparms.h enum DistributionType
}
