% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-params.R
\name{ElementParams}
\alias{ElementParams}
\title{Element Parameters (opaque)}
\usage{
ElementParams(ptr = NULL)
}
\arguments{
\item{ptr}{External pointer (internal use)}
}
\description{
Wraps \verb{std::shared_ptr<typename DCRTPoly::Params>} on the C++ side.
Used by the \code{params} argument of CKKS plaintext factories and
returned by \code{get_element_params()}. This class ships as
scaffolding only: no constructor surface other than
wrapping an existing external pointer.
}
