% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ckks.R
\name{eval_bootstrap_key_gen}
\alias{eval_bootstrap_key_gen}
\title{Generate bootstrapping keys}
\usage{
eval_bootstrap_key_gen(cc, sk, slots)
}
\arguments{
\item{cc}{A CryptoContext}

\item{sk}{A PrivateKey}

\item{slots}{Number of slots}
}
\description{
Generate bootstrapping keys
}
