% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ckks.R
\name{eval_cos}
\alias{eval_cos}
\title{Evaluate cosine on a ciphertext}
\usage{
eval_cos(ct, a, b, degree)
}
\arguments{
\item{ct}{A Ciphertext}

\item{a}{Lower bound of the approximation interval}

\item{b}{Upper bound of the approximation interval}

\item{degree}{Chebyshev polynomial degree}
}
\description{
Evaluate cosine on a ciphertext
}
