% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ckks.R
\name{eval_logistic}
\alias{eval_logistic}
\title{Evaluate logistic function on a ciphertext}
\usage{
eval_logistic(ct, a, b, degree)
}
\arguments{
\item{ct}{A Ciphertext}

\item{a}{Lower bound of the approximation interval}

\item{b}{Upper bound of the approximation interval}

\item{degree}{Chebyshev polynomial degree}
}
\description{
Evaluate logistic function on a ciphertext
}
