% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiparty.R
\name{EvalKey}
\alias{EvalKey}
\title{EvalKey class for multi-party key operations}
\usage{
EvalKey(ptr = NULL)
}
\arguments{
\item{ptr}{External pointer (internal use)}
}
\description{
EvalKey class for multi-party key operations
}
