% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{ExecutionMode}
\alias{ExecutionMode}
\title{Execution Mode
Source: pke/constants-defs.h enum ExecutionMode}
\usage{
ExecutionMode
}
\description{
Execution Mode
Source: pke/constants-defs.h enum ExecutionMode
}
