% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{Feature}
\alias{Feature}
\title{PKE Scheme Features (bitmask)
Source: pke/constants-defs.h enum PKESchemeFeature}
\usage{
Feature
}
\description{
PKE Scheme Features (bitmask)
Source: pke/constants-defs.h enum PKESchemeFeature
}
