% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{key_gen}
\alias{key_gen}
\title{Generate key pair}
\usage{
key_gen(cc, ...)
}
\arguments{
\item{cc}{A CryptoContext}

\item{...}{Method-specific arguments (eval_mult, rotations)}
}
\value{
A KeyPair
}
\description{
Generate key pair
}
