% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{KeygenMode}
\alias{KeygenMode}
\title{Key Generation Mode
Source: binfhe/binfhe-constants.h enum KEYGEN_MODE}
\usage{
KeygenMode
}
\description{
Key Generation Mode
Source: binfhe/binfhe-constants.h enum KEYGEN_MODE
}
