% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{KeyPair}
\alias{KeyPair}
\title{Key Pair}
\usage{
KeyPair(public = NULL, secret = NULL)
}
\arguments{
\item{public}{A PublicKey}

\item{secret}{A PrivateKey}
}
\description{
Contains a public key and a secret (private) key.
}
