% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{KeySwitchTechnique}
\alias{KeySwitchTechnique}
\title{Key Switching Techniques
Source: pke/constants-defs.h enum KeySwitchTechnique}
\usage{
KeySwitchTechnique
}
\description{
Key Switching Techniques
Source: pke/constants-defs.h enum KeySwitchTechnique
}
