% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binfhe.R
\name{LWECiphertext}
\alias{LWECiphertext}
\title{LWE Ciphertext (Binary FHE)}
\usage{
LWECiphertext(ptr = NULL)
}
\arguments{
\item{ptr}{External pointer (internal use)}
}
\description{
LWE Ciphertext (Binary FHE)
}
