% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binfhe.R
\name{LWEPrivateKey}
\alias{LWEPrivateKey}
\title{LWE Private Key (Binary FHE)}
\usage{
LWEPrivateKey(ptr = NULL)
}
\arguments{
\item{ptr}{External pointer (internal use)}
}
\description{
LWE Private Key (Binary FHE)
}
