% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{MultipartyMode}
\alias{MultipartyMode}
\title{Multiparty Mode
Source: pke/constants-defs.h enum MultipartyMode}
\usage{
MultipartyMode
}
\description{
Multiparty Mode
Source: pke/constants-defs.h enum MultipartyMode
}
