% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{MultiplicationTechnique}
\alias{MultiplicationTechnique}
\title{Multiplication Technique (BFV)
Source: pke/constants-defs.h enum MultiplicationTechnique}
\usage{
MultiplicationTechnique
}
\description{
Multiplication Technique (BFV)
Source: pke/constants-defs.h enum MultiplicationTechnique
}
