% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfhe-package.R
\docType{package}
\name{openfhe.R-package}
\alias{openfhe.R}
\alias{openfhe.R-package}
\title{openfhe.R: R Interface to the OpenFHE Fully Homomorphic Encryption Library}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides an R interface to 'OpenFHE', the open-source C++ library for fully homomorphic encryption (Al Badawi and others, 2022) \url{https://eprint.iacr.org/2022/915}, which allows computation directly on encrypted data without access to the secret key. Supports the Brakerski-Fan-Vercauteren (BFV, 2012) \url{https://eprint.iacr.org/2012/144}, Brakerski-Gentry-Vaikuntanathan (BGV, 2014) \doi{10.1145/2633600}, and Cheon-Kim-Kim-Song (CKKS, 2017) \url{https://eprint.iacr.org/2016/421} schemes for arithmetic on encrypted numbers, together with the Ducas-Micciancio (FHEW, 2015) \url{https://eprint.iacr.org/2014/816} and Chillotti-Gama-Georgieva-Izabachene (TFHE, 2020) \url{https://eprint.iacr.org/2018/421} schemes for evaluating arbitrary functions on encrypted bits.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://openfheorg.github.io/openfhe.R/}
  \item \url{https://github.com/openfheorg/openfhe.R}
  \item Report bugs at \url{https://github.com/openfheorg/openfhe.R/issues}
}

}
\author{
\strong{Maintainer}: Balasubramanian Narasimhan \email{naras@stanford.edu} (\href{https://orcid.org/0000-0001-5852-7639}{ORCID})

Authors:
\itemize{
  \item Balasubramanian Narasimhan \email{naras@stanford.edu} (\href{https://orcid.org/0000-0001-5852-7639}{ORCID})
}

Other contributors:
\itemize{
  \item Authors of the OpenFHE C++ library [contributor, copyright holder]
}

}
