% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfhe-object.R
\name{OpenFHEObject}
\alias{OpenFHEObject}
\title{Base class for OpenFHE objects}
\usage{
OpenFHEObject(ptr = NULL)
}
\arguments{
\item{ptr}{External pointer to C++ object (internal use)}
}
\description{
All OpenFHE objects (CryptoContext, Ciphertext, Plaintext, Keys, etc.)
inherit from this class. It holds an external pointer to a C++ shared_ptr.
}
