% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{PlaintextEncodings}
\alias{PlaintextEncodings}
\title{Plaintext Encoding Types
Source: pke/constants-defs.h enum PlaintextEncodings}
\usage{
PlaintextEncodings
}
\description{
Plaintext Encoding Types
Source: pke/constants-defs.h enum PlaintextEncodings
}
