% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{ScalingTechnique}
\alias{ScalingTechnique}
\title{Scaling Techniques (CKKS)
Source: pke/constants-defs.h enum ScalingTechnique}
\usage{
ScalingTechnique
}
\description{
Scaling Techniques (CKKS)
Source: pke/constants-defs.h enum ScalingTechnique
}
