% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{SecretKeyDist}
\alias{SecretKeyDist}
\title{Secret Key Distribution
Source: core/lattice/constants-lattice.h enum SecretKeyDist}
\usage{
SecretKeyDist
}
\description{
Secret Key Distribution
Source: core/lattice/constants-lattice.h enum SecretKeyDist
}
