% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaintext.R
\name{set_length}
\alias{set_length}
\title{Set the effective length of a plaintext}
\usage{
set_length(pt, len)
}
\arguments{
\item{pt}{A Plaintext}

\item{len}{Integer length}
}
\description{
Set the effective length of a plaintext
}
